Created initial scaffolding.
The spec hasn't been changed (there are going to be some additions soon-ish),
but I've started work on making this a reality.

Anyways, right now, there's a mysql schema (only for mysql) that you can use
to insert into a database and have the basis of what's needed for Branch.

Change might be made to the schema, so, beware...

I'm attempting to use object oriented php as opposed to my previous PHP
excursions, which should hopefully make things better.

A future hope of mine is that by the end, the entire database will be
abstracted away and the entity classes will perform most of those actions. I
really like how Django and Rails do this.