More Updates!
So you can now create a user in the database.

The Users class is mostly static, being that I want one to call `Users::foo`
to do things. There's a DBUser class which will be the object that gets
manipulated by people (as in DBUser->changePassword, etc). Users is just for
global user-y things (for now).

Passwords are salted and hashed (hashed a lot). So guessing a password and
attempting to brute force your way in will be a little hard.

The DB abstraction layer, BranchDB (okay, not all that abstracted, but a little
abstracted) automatically processes queries and safe-ificates them. Which is
good.

There isn't anything else it can do besides create a user, but that's enough
for now, right?