Changed how login works, started on a playground
The old Users::login function is now Users::isPasswordCorrect (okay, I'm
probably going to change this to Users::validatePassword, but I didn't think
of that name at the time when I changed it). The new Users::login is the start
of a framework to generate a session that declares that you're logged in.

Speaking of sessions, while I haven't started working on that yet, they're
probably going to be their own class, just to make things easier on myself (I
hope?). I'm not quite sure yet...

So, don't actually use Users::login right now, just use
Users::isPasswordCorrect (that should probably be private, but, I'll get
there).

As for the playground, there are two new files called playground.php and
ajax.php. playground.php is the page for actually playing around in the
browser to see if things work. Right now it has two forms: a create user form
(which works); and a log user in (which worked, until the Users::login change,
which I'm currently working on).

ajax.php is the page that ajax request are made to from the playground. One of
the big Ghost features that I really like was the previewing everything before
performing an action, which to me says ajax and potentially pushState (but
mostly ajax to get data and such). I think that having a page specifically for
processing requests is okay.

It isn't secure though (yet?). As a part of the session stuff (which is what
I'm going to work on next) is securing the site as much as possible. After
sessions are put together, I'm going to work on having playback require a
login (or, have a variable switch, I'll build an installer later...) to access
it.