More playground stuff.
Sessions are now a thing. They're a fairly simple abstraction layer for
$_SESSION, but, as your sterotypical programmer, I like abstractions...

This means that there is logging in actions to be had! Well, it kinda/sorta
requires that you have a user in the database. If you're going to try to run
this on your machine, you should comment out the if logged in parts of the
user creation, create a user, and take the comments back out.

User creatation will eventually have a security layer, most likely in where
it's called (well, actually, having it in the function itself means that if
plugins are eventually supported a plugin can't just call Users::create(...),
so, this is something I'm going to need to think about).

My next task is actually making posts. This'll be fun, right? Especially if
you consider how I'm going to store posts (each post has a file on the server
which has a copy of the database's contents in yaml (with tags expressed as an
array) and the content of the post itself in markdown). So, fun times are
ahead...