Hey, there are posts now, kinda...
All posts can be queried (but not specific ones, yet). The scaffolding for
getting specific posts is done, I just need to apply them...

Posts are done using a hybrid database/file mechanism. My thinking here is
that by putting the contents of the posts (and duplicating their database
information in yaml frontmatter) sites are infinitely more portable, post
contents are safter from sql injections (I'll probably explain that more
later).

Eventually and import tool will be built, but that's a task for much later.

I should add a pre-commit hook to export the database's sql for people to use
while in development...