More Work On Themes
I did a little more research on not sending a password in plaintext (which,
right now is what happens, but isn't secure in any manner). I updated the
passwords spec to detail what *will* happen in the future (the very near
future, as in, it's the next thing I'm going to do).

The playground is now empty (well, not in code) because I'm reworking it to
look all pretty like with the theme.