More Ajax and Playground fun
I started rebuilding playground functionality using the theme instead of an
ugly blank page. While the new page isn't all that much better, I'm working on
it, and hopefully it'll get prettier eventually...

Right now creating users and viewing all users work. Also changing
passwords...

That's right, passwords can be changed now. Which is important, because I
implemented the nonce/password sha'd combination for network requests, because
I really don't want to do all that much maintenance when this is finished
(security = important). For more information on the nonce system, see the
password spec.

The nonce system has also been implemented with all other times you send a
password over the network. This is a good thing.

Because of the need to `sha256` things using the client, the crypto-js sha256
module has been added to the project. Because the other modules aren't needed,
they haven't been added.

A new misc functions file has been added, right now it just generates random
strings (which is used to create a nonce), eventually it might have more
functions for use around Branch.

Which brings me to: themes should *not* have functions. There isn't any reason
for a theme that's tightly integrated with the system (as in, there is only
one theme ever, you may modify it as you please, but it's supposed to be an
overhaul, or dropping a new theme from someone else in the directory), so I'm
not going to follow WordPress here.

While I've never mentioned having a theme functions file before, I wrote that
more for myself, because I've been considering it.