It's now harder for people to read the config file
Well, not harder for the person writing it, or managing the system, but harder
for the person visiting the website in their browser.

I just wrapped the sucker in `<?php` and `?>`, have the `?>` commented out in
yaml form, and parse out the `<?php\n` in the file. And the config works.

Is it the fastest? Probably not. Because it's reading a file, but this isn't a
change from before. There are probably better ways to do it, well, more
efficient ways, but this way is nice and readable for the people using this...

And it isn't amazingly slow.