More branch -> small fixes
I still need to get a bunch of things changed, like the whole BranchDB
class... That should probably be SmallDB, or something similar.

Plus the CSS should be looked at, especially considering that it's going to
be the thing changed the most. It's not that I don't like the current style,
it's just that it isn't anywhere near what it should be in the near-future.