More branch -> small fixes (and more!)
I made a few more branch -> small changes (because I think that *maybe* the
project should use it's name, maybe).

I also added in a save method to DBUser, which means that they can be played
with, and their new data can be put in the database without extra queries. Up
next is a delete method for DBUser.

A new, not yet full integrated change is that users no longer have a username
(instead they'll sign in by email). Instead, the name field will be what each
user is referred to throughout (again, eventually). This change is being
implemented because email is unique enough, and if we wanted to add in OAuth
or OpenID login instead (hey, it could be a cool idea/thing to do at some
point) it becomes slightly easier (potentially, I don't actually know what's
involved in that process).