Some minor style work.
I made some tweaks to the style. I know that it's still premature for that,
but considering that the big draw should be that it looks pretty, it's still
fairly important.

I also put the new post creator back in the playground (I can't remember why
that went away...).

As for the style tweaks:

-   I removed everything from the old stylesheet (which is why the playground
    no longer has tabs, sorry). I did this because the old styles, while
    acceptable, don't really emphasize the text, which is the whole goal of
    the application. Plus there weren't all that many to begin with.

-   I added two new fonts (and got rid of one old): Bitter and Serif Beta.
    While Bitter isn't currently in use (it was at one point), I think it
    would be a nice title font if we don't want to use SerifBeta (the body
    text). I think either is acceptable.

At some point a monospaced font should be aquired (probably Cousine or Source
Code Pro, because they're both nice), but that's not important yet (neither
was the style tweaking).

The actual admin center is next, and along with it will be some new specs,
because the old ones haven't been updated with the new vision.