Hopefully entirely switched from username -> email
With small, I don't want there to be a username, I want users to just sign in
with their email address. Is it slightly harder for the user, maybe, but
that's not really my concern.

By using an email address instead of a username, it removes one unnecessary
field, the username one. Users instead should use the name field for their
display name, which is what they'll be referred to everywhere other than the
login section.

So... That's hopefully now switched over. If it isn't, I'll find it
eventually.

With the switch, a new schema is now in use. It's removed the tags table (or
renamed it) and added a categories table. All posts belong to one category,
that's just the way it is.

In other news, I forgot to remove the pages table, because they aren't a thing
anymore. So... I'll get to that at some point, probably the next schema
changes.