JavaScript Work. Mostly Login.
I'm starting to put multiuse JavaScript functions in one nice file (a .js.php
file, because it's cool...). Right now it's mostly login stuff canaballized
from the theme/header.php file.

Speaking of the header.php, it's probably going to go away, soon. I've created
a new head.php file which contains only the content inside of the <head> tag,
which leaves the other page the responsibility of everything else (for now,
there might be some page types in the /theme directory soonish).

As for the login stuff, /small/login.php works now. And it looks acceptable.
both of which are bonuses. This means I now have to work on the administrative
area (which was the goal anyways). If you're looking at the login page, as
soon as you change the email input (which means you have to tab to the
password field or deselect it, I may make it update on keystroke, which would
be cool, too) it'll change to your gravatar (or identicon based on your email
hash). So, that's kinda cool.

I also added Source Code Pro as a font. It, like Bitter, is currently unused.
Unlike Bitter, it'll probably be used sooner, with posts, and stuff.

Still haven't removed the pages stuff...