Played with Admin area again
The admin area is starting to be nice. By starting I mean the cool developer-y
things are looking nice, the actual interface is just a bar with changing
content depending on which pseudo-tab you click.

And on the pseudo-tab front, they work, again. And, thanks to history.pushState
and window.onpopstate the forward and backward buttons in your browser (and on
your five+ button mouse) also work inside of the admin interface. Which is nice,
because I don't want to maintain more than one page at this point.

They probably won't be used for the entire site, because actual pages make sense
for public facing content, but for the administrative section, one make sense to
me.

On the login section, while I really wanted to modularize as many components as
possible, that doesn't seem like it's going to happen, because of JavaScripts
gorram differed. I want blocking, I'm not going to get blocking. That's all...